adaptation des poids en nombre flotants

+++ weight.hpp
@@ -8,7 +8,7 @@
 constexpr size_t WEIGHTS_SIZE = 3;
 
 //TODO: peut être passer ça en std::vector ?
-using Weight = std::array<int, WEIGHTS_SIZE>;
+using Weight = std::array<float, WEIGHTS_SIZE>;
 using Weights = std::vector<Weight>;
 
 bool operator>(const Weight& w1, const Weight& w2)
